requirements-dev.txt
pandas
matplotlib.pyplot
seaborn 
numpy
yaml